pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "voicematch"
version = "0.1.0"
description = "Voice matching pipeline for AI avatars"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.5.0",
    "gradio>=4.0.0",
    "pydub>=0.25.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "types-requests>=2.31.0",
]
gpu = [
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "resemblyzer>=0.1.4",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "numpy>=1.24.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--import-mode=importlib",
    "-v",
    "--tb=short",
    "--cov=src/voicematch",
    "--cov-report=term-missing",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
# Relaxed type checking for MVP - enable strict mode in future
warn_return_any = false
warn_unused_ignores = true
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = true
